requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
click==8.2.1
discord.py==2.6.3
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
